tools/incident-crawler/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
PyYAML>=6.0
anthropic>=0.34.0
openai>=1.0.0
tweepy>=4.14.0
praw>=7.7.0
